requirements.txt
fastapi==0.110.0
httpx==0.27.2
uvicorn==0.27.1
python-dotenv==1.0.1
tiktoken==0.5.2
